demos/demo1-indirect-prompt-injection/pyproject.toml
[project]
name = "demo1-indirect-prompt-injection"
version = "0.1.0"
description = "Demo: Indirect Prompt Injection Attack on Offer Analysis Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flock-core",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
